+++ scripts/macros.js
@@ -99,10 +99,13 @@ export const resetBestiary = async () => {
     if(!confirmed) return;
 
     const bestiary = game.settings.get('pf2e-bestiary-tracking', 'bestiary-tracking');
-    for(var monsterKey in bestiary.monster){
-        const monster = bestiary.monster[monsterKey];
-        const journalEntry = game.journal.getName(bestiaryJournalEntry);
-        await journalEntry.pages.get(monster.system.details.playerNotes.document)?.delete();
+
+    const journalEntry = game.journal.getName(bestiaryJournalEntry);
+    if(journalEntry){
+        for(var monsterKey in bestiary.monster){
+            const monster = bestiary.monster[monsterKey];
+            await journalEntry.pages.get(monster.system.details.playerNotes.document)?.delete();
+        }
     }
 
     await game.settings.set('pf2e-bestiary-tracking', 'bestiary-tracking', {
